Moved sensitivity sim tests to a separate class.
